+++ libcircuit_reliability/src/global_functions_and_constants/terminal_color_constants.hpp
@@ -4,7 +4,8 @@
 // Terminal coloring
 static const std::string kEndColor = "\033[0m";
 static const std::string kRed = "\033[1;31m";
-static const std::string kGreen = "\033[1;32m"; 
+static const std::string kOrange = "\033[1;33m";
+static const std::string kGreen = "\033[1;32m";
 // static const std::string kYellow = "\033[1;33m";
 static const std::string kYellow = "\033[0;33m";
 static const std::string kBlue = "\033[1;34m";