+++ libcircuit_reliability/src/simulations_handling/spectre_simulations/standard_simulation.hpp
@@ -50,7 +50,7 @@ private:
 	 * @param scenario_id
 	 * @return spectre output
 	 */
-	int RunSpectre( std::string scenario_id );
+	int RunSpectre( const std::string& scenario_id );
 
 	/**
 	 * @brief Virtual overrides TestSetup